[Breacking] Removed the Off filter.
The Off filter has never been used and is problematic: some explicit logs
MUST be emitted (creation of DependentToken) for instance. The "Off"
has been replaced with "Fatal" (only Fatal are logged). This is easier
to understand.

This change enabled to increment/decrement depth in the lock
when opening a group: parallel logs' depth is now totally synchronized
with the LogTime, no depth discrepancies can now occur.